extract test for Backend.leaves for reuse
